requirements.txt
Flask==2.2.2
Werkzeug<3.0
openpyxl==3.1.5
openai>=1.10.0
python-dotenv>=1.0.1
pymongo>=4.8.0
